[refs #27 & #36] > Continued the work with the CSS / Bootstrap / @media queries to improve the look of the site across a variety of screen types/sizes:
Adjusted the widths of some of the images and Bootstrap grid columns to different %s to behave more reponsively.
Re-jigged the column widths and arrangement of the social links, and added different icons to appear when viewing on smaller screens.
Changed the h1 heading to h2 to allow a better fit on the smallest mobile screens.
Removed the "async" attribute from the script call that loads the main JS file - realised this was causing an issue with the burger menus not being initially hidden when viewing the site on my phone, although this was not apparent with the user-agent-spoofing.

> Added meta descriptions to each of the pages to hopefully help with Google search results. Was going to add meta keywords too but reading up a little about it I see these are very old hat and no longer have any influence with site rankings!

> Had a response from the 123-Reg tech support team regarding the problem experienced with the web.config file contents - turns out they had omitted to include a particular opening & closing tag. Google still reckons I don't have GZIP compression enabled, however! #fml